requirements.txt
fastapi[standard]==0.135.1
uvicorn==0.42.0
pydantic==2.12.5
SQLAlchemy==2.0.48
cryptography==46.0.5
pwdlib[argon2]==0.3.0
python-jose==3.5.0
python-magic==0.4.27
numpy==2.4.3
imageio==2.37.2
Pillow==12.1.1
rich==14.2.0